NET: remove invalid check for maxclients
We arelady checking maxclients inside SV_FindProxy(),
second time not only redundant it plain invalid.
For example if you set maxclients to 1 not a single client would
be able to connect. If you set it to 2 then only one client would
be able to connect and that not simply an error by 1 since
if both clients connecting at the same frame they both would be rejected
despite that fact that proxy have only those two clients to server and
maxclients set to 2.

The only purpose that check serve was a highly hypotetical case
when you change (rcon/console) maxclints to lower value and you had some slow connecting client
during that process, in that case yeah, it would help.
But there still could be already connected clients so it does not serve
that case so it quite useless. Not to mention changing proxy settins on the fly
is somewhat wrong.